feat(refresh-token): Enhace logic throughout services
